requirements.txt
flask>=3.1
flask-cors>=6.0
replicate>=1.0
pyngrok>=7.5
python-dotenv>=1.0
